refactor: monaco workers
